Fix wrong assertion and add conditional branches
On leaf removing, when the entries of leaf node is less than [m/2], we
need to borrow its sibling for balance of the tree. Then we need to
consider which sibling to be borrowed from. The method is to check the
binary search result of parent's keys. When it is not found, we have to
consider different situations rather than use a simple assert statement
which just ignores some other branches.

Signed-off-by: Leo Ma <begeekmyfriend@gmail.com>